Added dm notices

+++ src/services/Channel.ts
@@ -358,28 +358,58 @@ export const deleteServerChannel = async (
   return [channelId, null];
 };
 
-export const upsertChannelNotice = async (channelId: string, content: string) => {
+export const upsertChannelNotice = async (content: string, where: { channelId?: string, userId?: string }) => {
+
+  if (where.channelId && where.userId) {
+    return [null, generateError('Only one of channelId and userId can be provided.' as const)] as const;
+  };
+
+  if (!where.channelId && !where.userId) {
+    return [null, generateError('Either channelId or userId must be provided.' as const)] as const;
+  }
+
+
+
   const notice = await prisma.chatNotice.upsert({
-    where: { channelId },
-    create: { id: generateId(), channelId, content },
+    where: where as { userId: string } | { channelId: string },
+    create: { id: generateId(), ...where as { userId: string } | { channelId: string }, content },
     update: { content },
   });
 
   return [notice, null] as const;
 }
 
-export const deleteChannelNotice = async (channelId: string) => {
+export const deleteChannelNotice = async (where: { channelId?: string, userId?: string }) => {
+  if (where.channelId && where.userId) {
+    return [null, generateError('Only one of channelId and userId can be provided.' as const)] as const;
+  };
+
+  if (!where.channelId && !where.userId) {
+    return [null, generateError('Either channelId or userId must be provided.' as const)] as const;
+  }
+
+
+
   const res = await prisma.chatNotice.delete({
-    where: { channelId },
+    where: where as { userId: string } | { channelId: string },
     select: { id: true }
   }).catch(() => { });
   if (!res) return [null, generateError('Channel notice does not exist.' as const)] as const;
   return [true, null] as const;
 }
 
-export const getChannelNotice = async (channelId: string) => {
+export const getChannelNotice = async (where: { channelId?: string, userId?: string }) => {
+  if (where.channelId && where.userId) {
+    return [null, generateError('Only one of channelId and userId can be provided.' as const)] as const;
+  };
+
+  if (!where.channelId && !where.userId) {
+    return [null, generateError('Either channelId or userId must be provided.' as const)] as const;
+  }
+
+
   const res = await prisma.chatNotice.findUnique({
-    where: { channelId },
+    where: where as { userId: string } | { channelId: string },
     select: { content: true, updatedAt: true, channelId: true, userId: true }
   }).catch(() => { });
   if (!res) return [null, generateError('Channel notice does not exist.' as const)] as const;
